bots/pyproject.toml
[project]
name = "bots"
version = "0.1.0"
description = ""
authors = [
    {name = "TheAppleKingy",email = "pridanov.ali.job@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram (>=3.21.0,<4.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "hupper (>=1.12.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "redis (>=6.2.0,<7.0.0)",
    "pydantic (>=2.11.7,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
